+++ include/cppjieba/SegmentBase.hpp
@@ -1,7 +1,7 @@
 #ifndef CPPJIEBA_SEGMENTBASE_H
 #define CPPJIEBA_SEGMENTBASE_H
 
-#include "limonp/Logging.hpp"
+#include "Utils.hpp"
 #include "PreFilter.hpp"
 #include <cassert>
 
@@ -9,8 +9,6 @@
 namespace cppjieba {
 
 const char* const SPECIAL_SEPARATORS = " \t\n\xEF\xBC\x8C\xE3\x80\x82";
-
-using namespace limonp;
 
 class SegmentBase {
  public: