+++ include/cppjieba/TextRankExtractor.hpp
@@ -2,11 +2,11 @@
 #define CPPJIEBA_TEXTRANK_EXTRACTOR_H
 
 #include <cmath>
+#include <map>
 #include "Jieba.hpp"
 #include "UnicodeFile.hpp"
 
 namespace cppjieba {
-  using namespace limonp;
   using namespace std;
 
   class TextRankExtractor {