+++ include/cppjieba/Unicode.hpp
@@ -7,7 +7,7 @@
 #include <string>
 #include <vector>
 #include <ostream>
-#include "limonp/LocalVector.hpp"
+#include "Utils.hpp"
 
 namespace cppjieba {
 
@@ -53,7 +53,7 @@ inline std::ostream& operator << (std::ostream& os, const RuneStr& r) {
   return os << "{\"rune\": \"" << r.rune << "\", \"offset\": " << r.offset << ", \"len\": " << r.len << "}";
 }
 
-typedef limonp::LocalVector<Rune> Unicode;
+typedef LocalVector<Rune> Unicode;
 typedef std::vector<RuneStr> RuneStrArray;
 
 // [left, right]